fixed get turn to return whos turn, fixed do turn to do a turn and got rid of function that did nothing. View needs to know that return of +.5 means theres a split.

+++ CS485Blackjack/CS485Blackjack/BlackjackPresenter.cpp
@@ -42,9 +42,9 @@ void BlackjackPresenter::addBet(int seat, Money cBank)
 	mcBJModel.addBet(seat, cBank);
 }
 
-void BlackjackPresenter::doTurn(int seat, int move, float hands)
+void BlackjackPresenter::doTurn(float seat, int move, float hands)
 {
-	mcBJModel.doTurn(seat, move, hands);
+	mcBJModel.doTurn(static_cast<int>(seat), move, hands);
 }
 
 float BlackjackPresenter::getTurn()
