+++ CS485Blackjack/CS485Blackjack/BlackjackPresenter.h
@@ -26,7 +26,7 @@ public:
 
 	void addBet(int seat, Money cBank);
 
-	void doTurn(int seat, int move, float hands);
+	void doTurn(float seat, int move, float hands);
 	float getTurn();
 	Card getCard();
 	void resetGame();
